Moved files to a new package
Made a new package called events and moved event-related files into package.
