+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.CSCB07Project"
         tools:targetApi="31">
         <activity
-            android:name=".EventDetails"
+            android:name=".events.EventDetails"
             android:exported="false" />
         <activity
-            android:name=".EventList"
+            android:name=".events.EventList"
             android:exported="false" />
         <activity
-            android:name=".AddEvent"
+            android:name=".events.AddEvent"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
