+++ app/src/main/java/com/example/cscb07_project/events/AddEvent.java
@@ -1,7 +1,6 @@
-package com.example.cscb07_project;
+package com.example.cscb07_project.events;
 import android.content.Intent;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -12,6 +11,7 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 
 
+import com.example.cscb07_project.R;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
