+++ app/src/main/java/com/example/cscb07_project/events/Event.java
@@ -1,4 +1,4 @@
-package com.example.cscb07_project;
+package com.example.cscb07_project.events;
 
 /**
  * Event model class.
