+++ app/src/main/java/com/example/cscb07_project/events/EventList.java
@@ -1,4 +1,4 @@
-package com.example.cscb07_project;
+package com.example.cscb07_project.events;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -11,6 +11,8 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
+import com.example.cscb07_project.MainActivity;
+import com.example.cscb07_project.R;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
