+++ app/src/main/res/layout/activity_add_new_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddEvent">
+    tools:context=".events.AddEvent">
 
     <EditText
         android:id="@+id/event_detail_name"
