+++ app/src/main/res/layout/event_item.xml
@@ -32,7 +32,7 @@
             android:layout_marginLeft="2dp"/>
 
         <Button
-            android:id="@+id/delete_button"
+            android:id="@+id/delete_event_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="delete"
